Harden worker crash recovery internals
