+++ crates/karva_ipc/src/protocol.rs
@@ -1,5 +1,7 @@
 //! Wire protocol shared by the Karva controller and worker processes.
 
+use std::sync::Arc;
+
 use karva_diagnostic::{RenderedDiagnostic, TestCaseResult};
 use karva_python_semantic::TestCacheKey;
 use serde::{Deserialize, Serialize};
@@ -7,22 +9,6 @@ use serde::{Deserialize, Serialize};
 /// One runtime state change sent from a worker to the controller.
 #[derive(Serialize, Deserialize)]
 pub enum WorkerEvent {
-    /// Worker entered one test's setup-or-execution lifecycle.
-    ///
-    /// The first checkpoint precedes fixture setup. A later checkpoint may
-    /// refine the display name after fixture-derived parameters resolve.
-    TestStarted {
-        /// Best display identity known when it differs from the stable key.
-        ///
-        /// Plain function names are omitted to avoid sending the same string
-        /// twice; the controller reconstructs them from `cache_key`.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        name: Option<String>,
-
-        /// Stable case identity used for completion and crash recovery.
-        cache_key: TestCacheKey,
-    },
-
     /// Test exceeded the configured slow-test threshold.
     TestSlow,
 
@@ -57,6 +43,27 @@ pub enum WireMessage {
     /// Selection and resume state sent by the controller after authentication.
     TestSelection(WorkerSelection),
 
+    /// Crash-durable identity for a test entering setup or execution.
+    ///
+    /// This direct frame avoids allocating and decoding a dispatch event that
+    /// never leaves the connection reader. A later frame may refine the
+    /// display identity after fixture-derived parameters resolve. Its compact
+    /// serialized keys limit traffic from one durable frame per test case;
+    /// source-level field names remain descriptive.
+    #[serde(rename = "C")]
+    TestCheckpoint {
+        /// Rendered parameter list without its function name or parentheses.
+        ///
+        /// The controller reconstructs the display identity from this suffix
+        /// and `cache_key`, avoiding duplicate function names on the wire.
+        #[serde(rename = "p", default, skip_serializing_if = "Option::is_none")]
+        parameters: Option<String>,
+
+        /// Stable case identity used for completion and crash recovery.
+        #[serde(rename = "k")]
+        cache_key: TestCacheKey,
+    },
+
     /// Runtime event sent from a worker to its controller.
     Event(Box<WorkerEvent>),
 }
@@ -64,8 +71,8 @@ pub enum WireMessage {
 /// Work owned by one worker generation.
 #[derive(Serialize, Deserialize)]
 pub struct WorkerSelection {
-    /// Exact test selectors in execution order.
-    pub test_paths: Vec<String>,
+    /// Exact test selectors in execution order, shared with controller recovery state.
+    pub test_paths: Vec<Arc<str>>,
 
     /// Runtime-expanded cases already completed by an earlier generation.
     pub resume_skip: Vec<TestCacheKey>,
