+++ crates/karva_ipc/src/worker.rs
@@ -4,6 +4,7 @@
 //! checkpoints that must survive a worker crash flush synchronously; ordinary
 //! events are coalesced for the configured interval.
 
+use std::fmt;
 use std::io::{BufReader, BufWriter, Write};
 use std::net::{Shutdown, SocketAddr, TcpStream};
 use std::sync::atomic::{AtomicBool, Ordering};
@@ -12,6 +13,8 @@ use std::thread::{self, JoinHandle};
 use std::time::Duration;
 
 use anyhow::{Context, Result, bail};
+use karva_python_semantic::{QualifiedFunctionName, QualifiedTestName};
+use serde::{Serialize, Serializer};
 
 use crate::protocol::{WireMessage, WorkerEvent, WorkerSelection};
 
@@ -41,6 +44,50 @@ struct WorkerConnection {
 
     /// Join handle retained so completion can report a flusher panic.
     flusher: Mutex<Option<JoinHandle<()>>>,
+}
+
+/// Allocation-free checkpoint frame matching [`WireMessage::TestCheckpoint`].
+#[derive(Serialize)]
+enum BorrowedCheckpointFrame<'a> {
+    /// Active test identity borrowed from the execution reporter.
+    #[serde(rename = "C")]
+    TestCheckpoint {
+        /// Rendered parameter list without its function name or parentheses.
+        #[serde(rename = "p", skip_serializing_if = "Option::is_none")]
+        parameters: Option<&'a str>,
+
+        /// Stable function and case identity serialized as one cache key.
+        #[serde(rename = "k")]
+        cache_key: BorrowedTestCacheKey<'a>,
+    },
+}
+
+/// Stable cache identity serialized without constructing an owned string.
+struct BorrowedTestCacheKey<'a> {
+    /// Qualified function portion of the cache key.
+    function_name: &'a QualifiedFunctionName,
+
+    /// Optional terminal parameter-case index.
+    case_index: Option<usize>,
+}
+
+impl fmt::Display for BorrowedTestCacheKey<'_> {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.function_name.fmt(formatter)?;
+        if let Some(case_index) = self.case_index {
+            write!(formatter, "[{case_index}]")?;
+        }
+        Ok(())
+    }
+}
+
+impl Serialize for BorrowedTestCacheKey<'_> {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.collect_str(self)
+    }
 }
 
 impl WorkerClient {
@@ -85,14 +132,27 @@ impl WorkerClient {
         Ok((client, selection))
     }
 
-    /// Queues one state change, synchronously committing crash checkpoints.
+    /// Queues one ordinary runtime state change.
     pub fn send_event(&self, event: WorkerEvent) -> Result<()> {
-        let flush = matches!(&event, WorkerEvent::TestStarted { .. })
-            || matches!(
-                &event,
-                WorkerEvent::TestFinished { result, .. } if result.outcome().is_non_success()
-            );
+        let flush = matches!(
+            &event,
+            WorkerEvent::TestFinished { result, .. } if result.outcome().is_non_success()
+        );
         self.write(&WireMessage::Event(Box::new(event)), flush)
+    }
+
+    /// Commits the active test identity before its setup or body can terminate the worker.
+    pub fn checkpoint(&self, test_name: &QualifiedTestName) -> Result<()> {
+        self.write(
+            &BorrowedCheckpointFrame::TestCheckpoint {
+                parameters: test_name.parameters(),
+                cache_key: BorrowedTestCacheKey {
+                    function_name: test_name.function_name(),
+                    case_index: test_name.case_index(),
+                },
+            },
+            true,
+        )
     }
 
     /// Marks the worker complete and gracefully closes the connection.
@@ -125,7 +185,7 @@ impl WorkerClient {
             .context("failed to close Karva controller connection")
     }
 
-    fn write(&self, message: &WireMessage, flush: bool) -> Result<()> {
+    fn write(&self, message: &impl Serialize, flush: bool) -> Result<()> {
         self.check_flush_error()?;
         let mut writer = self
             .connection
@@ -193,7 +253,7 @@ fn read_test_selection(stream: TcpStream) -> Result<WorkerSelection> {
     };
     match message.context("failed to read Karva worker test paths")? {
         WireMessage::TestSelection(selection) => Ok(selection),
-        WireMessage::Hello { .. } | WireMessage::Event(_) => {
+        WireMessage::Hello { .. } | WireMessage::TestCheckpoint { .. } | WireMessage::Event(_) => {
             bail!("Karva controller sent an invalid worker startup message")
         }
     }
@@ -274,5 +334,41 @@ impl Drop for WorkerConnection {
                 eprintln!("Karva worker event flusher panicked during cleanup: {error:?}");
             }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use karva_python_semantic::{ModulePath, QualifiedFunctionName, QualifiedTestName};
+
+    use super::{BorrowedCheckpointFrame, BorrowedTestCacheKey};
+    use crate::protocol::WireMessage;
+
+    #[test]
+    fn borrowed_checkpoint_matches_owned_wire_format() {
+        let test_name = QualifiedTestName::with_parameters(
+            QualifiedFunctionName::new(
+                "test_example".to_string(),
+                ModulePath::new_with_name("test.py", "tests.tést".to_string()),
+            ),
+            "value='é'".to_string(),
+        )
+        .with_case_index(Some(12));
+        let borrowed = BorrowedCheckpointFrame::TestCheckpoint {
+            parameters: test_name.parameters(),
+            cache_key: BorrowedTestCacheKey {
+                function_name: test_name.function_name(),
+                case_index: test_name.case_index(),
+            },
+        };
+        let owned = WireMessage::TestCheckpoint {
+            parameters: test_name.parameters().map(str::to_owned),
+            cache_key: test_name.cache_key(),
+        };
+
+        assert_eq!(
+            serde_json::to_value(borrowed).expect("serialize borrowed checkpoint"),
+            serde_json::to_value(owned).expect("serialize owned checkpoint")
+        );
     }
 }
