+++ crates/karva_runner/src/orchestration/streams.rs
@@ -3,7 +3,8 @@
 use std::fs::File;
 use std::io::{BufRead, BufReader, Write};
 use std::process::{ChildStderr, ChildStdout};
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::thread::{self, JoinHandle};
 
 // Receipt: the Python 3.14 abort diagnostic observed in CI was 5,848 bytes.
@@ -11,27 +12,29 @@ use std::thread::{self, JoinHandle};
 // grow the controller's per-worker spool without bound.
 const MAX_WORKER_STDERR_CAPTURE_BYTES: usize = 1024 * 1024;
 
-#[derive(Debug)]
 /// Background stdout drain that preserves worker output ordering.
+#[derive(Debug)]
 pub(super) struct WorkerOutputForwarder {
     /// Thread draining worker stdout in order.
     handle: JoinHandle<std::io::Result<()>>,
+
     /// Shared stop flag used after forced disconnect.
-    enabled: Arc<Mutex<bool>>,
+    enabled: Arc<AtomicBool>,
 }
 
-#[derive(Debug)]
 /// Background stderr drain that forwards output and keeps a bounded copy.
+#[derive(Debug)]
 pub(super) struct WorkerStderrForwarder {
     /// Thread forwarding and capturing worker stderr.
     handle: JoinHandle<std::io::Result<()>>,
+
     /// Shared stop flag used after forced disconnect.
-    enabled: Arc<Mutex<bool>>,
+    enabled: Arc<AtomicBool>,
 }
 
 impl WorkerStderrForwarder {
     pub(super) fn spawn(stderr: ChildStderr, captured: File) -> Self {
-        let enabled = Arc::new(Mutex::new(true));
+        let enabled = Arc::new(AtomicBool::new(true));
         let forwarder_enabled = Arc::clone(&enabled);
         let handle =
             thread::spawn(move || forward_worker_stderr(stderr, captured, &forwarder_enabled));
@@ -40,7 +43,7 @@ impl WorkerStderrForwarder {
 
     pub(super) fn join(self, worker_id: usize, wait: bool) {
         if !wait && !self.handle.is_finished() {
-            stop_forwarding(&self.enabled, worker_id, "stderr");
+            stop_forwarding(&self.enabled);
             return;
         }
         match self.handle.join() {
@@ -57,7 +60,7 @@ impl WorkerStderrForwarder {
 
 impl WorkerOutputForwarder {
     pub(super) fn spawn(stdout: ChildStdout) -> Self {
-        let enabled = Arc::new(Mutex::new(true));
+        let enabled = Arc::new(AtomicBool::new(true));
         let forwarder_enabled = Arc::clone(&enabled);
         let handle = thread::spawn(move || forward_worker_stdout(stdout, &forwarder_enabled));
         Self { handle, enabled }
@@ -65,7 +68,7 @@ impl WorkerOutputForwarder {
 
     pub(super) fn join(self, worker_id: usize, wait: bool) {
         if !wait && !self.handle.is_finished() {
-            stop_forwarding(&self.enabled, worker_id, "stdout");
+            stop_forwarding(&self.enabled);
             return;
         }
         match self.handle.join() {
@@ -81,16 +84,12 @@ impl WorkerOutputForwarder {
     }
 }
 
-fn stop_forwarding(enabled: &Mutex<bool>, worker_id: usize, stream: &str) {
-    match enabled.lock() {
-        Ok(mut enabled) => *enabled = false,
-        Err(error) => {
-            tracing::warn!(target: "karva_runner::orchestration", worker_id, "failed to stop worker {stream}: {error}");
-        }
-    }
+/// Stops future forwarding without waiting for an in-progress output write.
+fn stop_forwarding(enabled: &AtomicBool) {
+    enabled.store(false, Ordering::Relaxed);
 }
 
-fn forward_worker_stdout(stdout: ChildStdout, enabled: &Mutex<bool>) -> std::io::Result<()> {
+fn forward_worker_stdout(stdout: ChildStdout, enabled: &AtomicBool) -> std::io::Result<()> {
     let mut reader = BufReader::new(stdout);
     let mut line = Vec::new();
 
@@ -101,22 +100,32 @@ fn forward_worker_stdout(stdout: ChildStdout, enabled: &Mutex<bool>) -> std::io:
             return Ok(());
         }
 
-        let enabled = enabled.lock().map_err(|error| {
-            std::io::Error::other(format!("stdout forwarding lock poisoned: {error}"))
-        })?;
-        if !*enabled {
+        // Batch complete lines already buffered from the pipe. Retaining a
+        // partial final line prevents output from different workers interleaving.
+        loop {
+            let complete_bytes = {
+                let buffered = reader.fill_buf()?;
+                let Some(last_newline) = buffered.iter().rposition(|byte| *byte == b'\n') else {
+                    break;
+                };
+                line.extend_from_slice(&buffered[..=last_newline]);
+                last_newline + 1
+            };
+            reader.consume(complete_bytes);
+        }
+
+        if !enabled.load(Ordering::Relaxed) {
             return Ok(());
         }
         let mut stdout = std::io::stdout().lock();
         stdout.write_all(&line)?;
-        drop(enabled);
     }
 }
 
 fn forward_worker_stderr(
     stderr: ChildStderr,
     mut captured: File,
-    enabled: &Mutex<bool>,
+    enabled: &AtomicBool,
 ) -> std::io::Result<()> {
     let mut reader = BufReader::new(stderr);
     let mut capture_error = None;
@@ -130,10 +139,7 @@ fn forward_worker_stderr(
             break;
         }
         let consumed = bytes.len();
-        let enabled = enabled.lock().map_err(|error| {
-            std::io::Error::other(format!("stderr forwarding lock poisoned: {error}"))
-        })?;
-        if !*enabled {
+        if !enabled.load(Ordering::Relaxed) {
             return Ok(());
         }
         received_bytes = received_bytes.saturating_add(consumed);
@@ -153,7 +159,6 @@ fn forward_worker_stderr(
             forward_error = Some(error);
         }
         reader.consume(consumed);
-        drop(enabled);
     }
     if last_byte.is_some_and(|byte| byte != b'\n')
         && forward_error.is_none()
