+++ crates/karva_runner/src/partition/tests/mod.rs
@@ -1,0 +1,7 @@
+//! Partition behavior tests grouped by policy and recovery responsibility.
+
+mod balancing;
+mod helpers;
+mod ordering;
+mod recovery;
+mod selection;
